+++ fracmanager/cache_maintainer.go
@@ -144,7 +144,7 @@ func (cm *CacheMaintainer) CreateSortDocsCache() *cache.Cache[[]byte] {
 func (cm *CacheMaintainer) CreateIndexCache() *frac.IndexCache {
 	return &frac.IndexCache{
 		MIDs:       newCache[[]byte](cm, midsName),
-		RIDs:       newCache[[]byte](cm, ridsName),
+		RIDs:       newCache[[]uint64](cm, ridsName),
 		Params:     newCache[seqids.BlockParams](cm, paramsName),
 		LIDs:       newCache[*lids.Block](cm, lidsName),
 		Tokens:     newCache[*token.Block](cm, tokensName),
